Move string methods to calls_strings.py (#418)
Moves 20 methods to the new CallsStringsMixin:
- Basic: string_length, string_concat, string_slice,
  string_char_code, string_from_char_code, string_repeat
- Search: string_contains, string_starts_with, string_ends_with,
  string_index_of
- Transform: string_upper, string_lower, string_replace,
  string_split, string_join, string_strip
- To-string conversions: to_string (Int), bool_to_string,
  byte_to_string, float_to_string

calls.py: 3,135 -> 565 lines (-2,570)
calls_strings.py: 23 -> 2,588 lines

calls.py is now the target size — just the dispatcher, type
inference helpers, and _infer_concat_elem_type.
